Adjust testing for ReadAudioFile

+++ tests/ReadAudioFile.cpp
@@ -5,11 +5,12 @@
 
 int main()
 {
+    // sample rate = 132000, duration = 5s
     const auto filePath = "files/Sine_wave_440.wav";
     const auto audioBuffer = shone::core::AudioBuffer{filePath};
     
     assert(audioBuffer.filePath() == filePath);
     assert(audioBuffer.sampleRate() == 132000);
     assert(audioBuffer.audioFrames().size() == 330000);
-    assert(audioBuffer.audioFrames().size() / static_cast<double>(audioBuffer.sampleRate()) == 5.0f);
+    assert(static_cast<int>(audioBuffer.audioFrames().size() / static_cast<double>(audioBuffer.sampleRate())) == 5);
 }
